优化软键盘

+++ android/src/main/kotlin/br/com/keyboard_utils/keyboard/KeyboardOptions.kt
@@ -5,7 +5,7 @@ import org.json.JSONObject
 /**
  * Created by Wilson Martins on 2019-10-25.
  */
-class KeyboardOptions(private val isKeyboardOpen: Boolean, private val height: Int) {
+class KeyboardOptions(private val isKeyboardOpen: Boolean, private val height: Float) {
     fun toJson(): String = JSONObject().apply {
         put("isKeyboardOpen", isKeyboardOpen)
         put("keyboardHeight", height)
